Move test code to correct file in `api` package
